+++ test/test_espeak.py
@@ -14,54 +14,39 @@
 # along with phonemizer. If not, see <http://www.gnu.org/licenses/>.
 """Test of the espeak backend"""
 
+# pylint: disable=missing-docstring
+# pylint: disable=redefined-outer-name
 
-import shutil
 import os
-import re
+import shutil
 import pytest
 
-import phonemizer.separator as separator
-from phonemizer.backend import EspeakBackend, EspeakMbrolaBackend
-from phonemizer.separator import Separator
-
-
-@pytest.mark.parametrize(
-    'version, expected',
-    [('eSpeak text-to-speech: 1.48.03 04.Mar.14 Data at:'
-      '/usr/lib/x86_64-linux-gnu/espeak-data', '1.48.03'),
-     ('speak text-to-speech: 1.48.03 04.Mar.14 Data at: /usr/local/Cellar/'
-      'espeak/1.48.04_1/share/espeak-dat', '1.48.03'),
-     ('eSpeak NG text-to-speech: 1.49.2  Data at: /espeak-ng-data', '1.49.2'),
-     ('eSpeak NG text-to-speech: 1.51-dev  '
-      'Data at: /share/espeak-ng-data', '1.51-dev'),
-     ('eSpeak NG text-to-speech: 1.51.1.2.3-dev '
-      'Data at: /share/espeak-ng-data', '1.51.1.2.3-dev')])
-def test_versions(version, expected):
-    found = re.match(EspeakBackend.espeak_version_re, version).group(1)
-    assert found == expected
+from phonemizer.backend import EspeakBackend
+from phonemizer.backend.espeak.wrapper import EspeakWrapper
+from phonemizer.separator import Separator, default_separator
 
 
 def test_english():
     backend = EspeakBackend('en-us')
     text = u'hello world\ngoodbye\nthird line\nyet another'
-    out = backend.phonemize(text, separator.default_separator, True)
+    out = backend.phonemize(text, default_separator, True)
     assert out == u'həloʊ wɜːld\nɡʊdbaɪ\nθɜːd laɪn\njɛt ɐnʌðɚ'
 
 
 def test_stress():
     backend = EspeakBackend('en-us', with_stress=False)
-    assert u'həloʊ wɜːld' == backend.phonemize(
-        'hello world', separator.default_separator, True)
+    assert backend.phonemize(
+        'hello world', default_separator, True) == u'həloʊ wɜːld'
 
     backend = EspeakBackend('en-us', with_stress=True)
-    assert u'həlˈoʊ wˈɜːld' == backend.phonemize(
-        u'hello world', separator.default_separator, True)
+    assert backend.phonemize(
+        u'hello world', default_separator, True) == u'həlˈoʊ wˈɜːld'
 
 
 def test_french():
     backend = EspeakBackend('fr-fr')
     text = u'bonjour le monde'
-    sep = separator.Separator(word=';eword ', syllable=None, phone=' ')
+    sep = Separator(word=';eword ', syllable=None, phone=' ')
     expected = u'b ɔ̃ ʒ u ʁ ;eword l ə ;eword m ɔ̃ d ;eword '
     out = backend.phonemize(text, sep, False)
     assert out == expected
@@ -77,7 +62,7 @@ def test_french():
 def test_arabic():
     backend = EspeakBackend('ar')
     text = u'السلام عليكم'
-    sep = separator.Separator()
+    sep = Separator()
 
     # Arabic seems to have changed starting at espeak-ng-1.49.3
     if tuple(EspeakBackend.version().split('.')) >= ('1', '49', '3'):
@@ -88,19 +73,24 @@ def test_arabic():
     assert out == expected
 
 
-@pytest.mark.skipif(
-    not EspeakBackend.is_espeak_ng(),
-    reason='language switch only exists for espeak-ng')
-def test_language_switch():
-    text = [
+@pytest.fixture
+def langswitch_text():
+    return [
         "j'aime l'anglais",
         "j'aime le football",
         "football",
         "surtout le real madrid",
         "n'utilise pas google"]
 
+
+@pytest.mark.skipif(
+    not EspeakBackend.is_espeak_ng(),
+    reason='language switch only exists for espeak-ng')
+@pytest.mark.parametrize('njobs', [1, 3])
+def test_language_switch_keep_flags(caplog, langswitch_text, njobs):
     backend = EspeakBackend('fr-fr', language_switch='keep-flags')
-    out = backend.phonemize(text, separator.Separator(), True)
+    out = backend.phonemize(
+        langswitch_text, separator=Separator(), strip=True, njobs=njobs)
     assert out == [
         'ʒɛm lɑ̃ɡlɛ',
         'ʒɛm lə (en)fʊtbɔːl(fr)',
@@ -108,9 +98,24 @@ def test_language_switch():
         'syʁtu lə (en)ɹiəl(fr) madʁid',
         'nytiliz pa (en)ɡuːɡəl(fr)']
 
+    messages = [msg[2] for msg in caplog.record_tuples]
+    assert (
+        '4 utterances containing language switches on lines 2, 3, 4, 5'
+        in messages)
+    assert (
+        'language switch flags have been kept (applying "keep-flags" policy)'
+        in messages)
+
+
+@pytest.mark.skipif(
+    not EspeakBackend.is_espeak_ng(),
+    reason='language switch only exists for espeak-ng')
+@pytest.mark.parametrize('njobs', [1, 3])
+def test_language_switch_default(caplog, langswitch_text, njobs):
     # default behavior is to keep the flags
     backend = EspeakBackend('fr-fr')
-    out = backend.phonemize(text, separator.Separator(), True)
+    out = backend.phonemize(
+        langswitch_text, separator=Separator(), strip=True, njobs=njobs)
     assert out == [
         'ʒɛm lɑ̃ɡlɛ',
         'ʒɛm lə (en)fʊtbɔːl(fr)',
@@ -118,8 +123,23 @@ def test_language_switch():
         'syʁtu lə (en)ɹiəl(fr) madʁid',
         'nytiliz pa (en)ɡuːɡəl(fr)']
 
+    messages = [msg[2] for msg in caplog.record_tuples]
+    assert (
+        '4 utterances containing language switches on lines 2, 3, 4, 5'
+        in messages)
+    assert (
+        'language switch flags have been kept (applying "keep-flags" policy)'
+        in messages)
+
+
+@pytest.mark.skipif(
+    not EspeakBackend.is_espeak_ng(),
+    reason='language switch only exists for espeak-ng')
+@pytest.mark.parametrize('njobs', [1, 3])
+def test_language_switch_remove_flags(caplog, langswitch_text, njobs):
     backend = EspeakBackend('fr-fr', language_switch='remove-flags')
-    out = backend.phonemize(text, separator.Separator(), True)
+    out = backend.phonemize(
+        langswitch_text, separator=Separator(), strip=True, njobs=njobs)
     assert out == [
         'ʒɛm lɑ̃ɡlɛ',
         'ʒɛm lə fʊtbɔːl',
@@ -127,9 +147,31 @@ def test_language_switch():
         'syʁtu lə ɹiəl madʁid',
         'nytiliz pa ɡuːɡəl']
 
+    messages = [msg[2] for msg in caplog.record_tuples]
+    assert (
+        '4 utterances containing language switches on lines 2, 3, 4, 5'
+        in messages)
+    assert (
+        'language switch flags have been removed '
+        '(applying "remove-flags" policy)'
+        in messages)
+
+
+@pytest.mark.skipif(
+    not EspeakBackend.is_espeak_ng(),
+    reason='language switch only exists for espeak-ng')
+@pytest.mark.parametrize('njobs', [1, 3])
+def test_language_switch_remove_utterance(caplog, langswitch_text, njobs):
     backend = EspeakBackend('fr-fr', language_switch='remove-utterance')
-    out = backend.phonemize(text, separator.Separator(), True)
+    out = backend.phonemize(
+        langswitch_text, separator=Separator(), strip=True, njobs=njobs)
     assert out == ['ʒɛm lɑ̃ɡlɛ']
+
+    messages = [msg[2] for msg in caplog.record_tuples]
+    assert (
+        'removed 4 utterances containing language switches '
+        '(applying "remove-utterance" policy)'
+        in messages)
 
     with pytest.raises(RuntimeError):
         backend = EspeakBackend('fr-fr', language_switch='foo')
@@ -144,10 +186,10 @@ def test_language_switch():
         'a comma, , a point.',
         'a comma? a point!']
      for s in (True, False)
-     for u in (separator.Separator(), separator.Separator(word='_', phone=' '))
+     for u in (Separator(), Separator(word='_', phone=' '))
      ))
 def test_punctuation(text, strip, sep):
-    if sep == separator.Separator():
+    if sep == Separator():
         expected = 'ɐ kɑːmə ɐ pɔɪnt' if strip else 'ɐ kɑːmə ɐ pɔɪnt '
     else:
         expected = (
@@ -160,7 +202,7 @@ def test_punctuation(text, strip, sep):
 # see https://github.com/bootphon/phonemizer/issues/31
 def test_phone_separator_simple():
     text = 'The lion and the tiger ran'
-    sep = separator.Separator(phone='_')
+    sep = Separator(phone='_')
     backend = EspeakBackend('en-us')
 
     output = backend.phonemize(text, separator=sep, strip=True)
@@ -179,7 +221,7 @@ def test_phone_separator_simple():
      ('He was hungry and tired.', 'h_iː w_ʌ_z h_ʌ_ŋ_ɡ_ɹ_i æ_n_d t_aɪɚ_d'),
      ('He was hungry but tired.', 'h_iː w_ʌ_z h_ʌ_ŋ_ɡ_ɹ_i b_ʌ_t t_aɪɚ_d')))
 def test_phone_separator(text, expected):
-    sep = separator.Separator(phone='_')
+    sep = Separator(phone='_')
     backend = EspeakBackend('en-us')
     output = backend.phonemize(text, separator=sep, strip=True)
     assert output == expected
@@ -189,42 +231,42 @@ def test_phone_separator(text, expected):
     'PHONEMIZER_ESPEAK_PATH' in os.environ,
     reason='cannot modify environment')
 def test_path_good():
-    espeak = EspeakBackend.espeak_path()
+    espeak = EspeakBackend.library()
     try:
-        EspeakBackend.set_espeak_path(None)
-        assert espeak == EspeakBackend.espeak_path()
+        EspeakBackend.set_library(None)
+        assert espeak == EspeakBackend.library()
 
-        binary = shutil.which('espeak')
-        EspeakBackend.set_espeak_path(binary)
+        library = EspeakWrapper().library_path
+        EspeakBackend.set_library(library)
 
         test_english()
 
     # restore the espeak path to default
     finally:
-        EspeakBackend.set_espeak_path(espeak)
+        EspeakBackend.set_library(None)
 
 
 @pytest.mark.skipif(
     'PHONEMIZER_ESPEAK_PATH' in os.environ,
     reason='cannot modify environment')
 def test_path_bad():
-    espeak = EspeakBackend.espeak_path()
     try:
         # corrupt the default espeak path, try to use python executable instead
         binary = shutil.which('python')
-        EspeakBackend.set_espeak_path(binary)
+        EspeakBackend.set_library(binary)
 
         with pytest.raises(RuntimeError):
-            EspeakBackend('en-us').phonemize('hello')
+            EspeakBackend('en-us')
         with pytest.raises(RuntimeError):
             EspeakBackend.version()
 
-        with pytest.raises(ValueError):
-            EspeakBackend.set_espeak_path(__file__)
+        EspeakBackend.set_library(__file__)
+        with pytest.raises(RuntimeError):
+            EspeakBackend('en-us')
 
     # restore the espeak path to default
     finally:
-        EspeakBackend.set_espeak_path(espeak)
+        EspeakBackend.set_library(None)
 
 
 @pytest.mark.skipif(
@@ -240,7 +282,7 @@ def test_path_venv():
             EspeakBackend.version()
 
         os.environ['PHONEMIZER_ESPEAK_PATH'] = __file__
-        with pytest.raises(ValueError):
+        with pytest.raises(RuntimeError):
             EspeakBackend.version()
 
     finally:
@@ -250,103 +292,15 @@ def test_path_venv():
             pass
 
 
-@pytest.mark.skipif(
-    not EspeakMbrolaBackend.is_available() or
-    not EspeakMbrolaBackend.is_supported_language('mb-fr1'),
-    reason='mbrola or mb-fr1 voice not installed')
-@pytest.mark.parametrize(
-    'text, expected',
-    [
-        # plosives
-        ('pont', 'po~'),
-        ('bon', 'bo~'),
-        ('temps', 'ta~'),
-        ('dans', 'da~'),
-        ('quand', 'ka~'),
-        ('gant', 'ga~'),
-        # fricatives
-        ('femme', 'fam'),
-        ('vent', 'va~'),
-        ('sans', 'sa~'),
-        ('champ', 'Sa~'),
-        ('gens', 'Za~'),
-        ('ion', 'jo~'),
-        # nasals
-        ('mont', 'mo~'),
-        ('nom', 'no~'),
-        ('oignon', 'onjo~'),
-        ('ping', 'piN'),
-        # liquid glides
-        ('long', 'lo~'),
-        ('rond', 'Ro~'),
-        ('coin', 'kwe~'),
-        ('juin', 'Zye~'),
-        ('pierre', 'pjER'),
-        # vowels
-        ('si', 'si'),
-        ('ses', 'se'),
-        ('seize', 'sEz'),
-        ('patte', 'pat'),
-        ('pâte', 'pat'),
-        ('comme', 'kOm'),
-        ('gros', 'gRo'),
-        ('doux', 'du'),
-        ('du', 'dy'),
-        ('deux', 'd2'),
-        ('neuf', 'n9f'),
-        ('justement', 'Zystma~'),
-        ('vin', 've~'),
-        ('vent', 'va~'),
-        ('bon', 'bo~'),
-        ('brun', 'bR9~')])
-def test_sampa_fr(text, expected):
-    assert expected == EspeakMbrolaBackend('mb-fr1').phonemize(
-
-        text, strip=True, separator=Separator(phone=''))
-
-
-@pytest.mark.skipif(
-    not EspeakMbrolaBackend.is_available() or
-    not EspeakMbrolaBackend.is_supported_language('mb-fr1'),
-    reason='mbrola or mb-fr1 voice not installed')
-def test_french_sampa():
-    text = u'bonjour le monde'
-    backend = EspeakMbrolaBackend('mb-fr1')
-    sep = separator.Separator(word=None, phone=' ')
-
-    expected = 'b o~ Z u R l @ m o~ d '
-    out = backend.phonemize(text, separator=sep, strip=False)
-    assert out == expected
-
-    expected = 'b o~ Z u R l @ m o~ d'
-    out = backend.phonemize(text, separator=sep, strip=True)
-    assert out == expected
-
-    assert '' == backend.phonemize('', separator=sep, strip=True)
-    assert '' == backend.phonemize('"', separator=sep, strip=True)
-
-
-@pytest.mark.skipif(
-    not EspeakMbrolaBackend.is_available(),
-    reason='mbrola not installed')
-def test_mbrola_bad_language():
-    assert not EspeakMbrolaBackend.is_supported_language('foo-bar')
-
-
-
-@pytest.mark.skipif(
-    EspeakBackend.version(as_tuple=True)[1] <= 48,
-    reason='tie option requires espeak>1.48')
 @pytest.mark.parametrize(
     'tie, expected', [
-        (False, 'dʒæki tʃæn '), (True, 'd͡ʒæki t͡ʃæn '), ('8', 'd8ʒæki t8ʃæn ')])
+        (False, 'dʒæki tʃæn '),
+        (True, 'd͡ʒæki t͡ʃæn '),
+        ('8', 'd8ʒæki t8ʃæn ')])
 def test_tie(tie, expected):
     assert EspeakBackend('en-us', tie=tie).phonemize('Jackie Chan') == expected
 
 
-@pytest.mark.skipif(
-    EspeakBackend.version(as_tuple=True)[1] <= 48,
-    reason='tie option requires espeak>1.48')
 def test_tie_bad():
     with pytest.raises(RuntimeError):
         EspeakBackend('en-us', tie='abc')